chore(app): integrate `useSession` in default layout and clean up old auth handling
Refactored default layout to use the `useSession` composable for session management directly, removing redundant authentication handling code. This change simplifies the sign-out logic and enhances the maintainability and readability of the code by eliminating the old, verbose sign-out function and directly binding the `session.signout` method to the sign-out event in the navigation bar. This commit also removes unnecessary error handling in the NuxtErrorBoundary, relying on the centralized error management within the app.