feat(app): add circular option and rename InputFile to InputImage
This update introduces a `circular` prop to the formerly named `InputFile` component, now renamed to `InputImage`. The `circular` prop allows the component to render image thumbnails in a circular shape, enhancing the visual options available to users. This change reflects the component's focus on image inputs rather than generic file inputs, hence the renaming to better align with its functionality. This enhancement aims to provide a more intuitive and visually appealing user experience when dealing with image uploads.